AuthenticationAssignment/Backend/requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
platformdirs==4.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.0
